+++ OOP/lab01/input.h
@@ -9,7 +9,7 @@
 
 return_code read_from_file(FILE *f, figure_t &fig);
 return_code read_line_point(FILE *f, point_t &p);
-return_code read_line_matrix(FILE *f, int &i, int &j);
+return_code read_line_matrix(FILE *f, size_t &i, size_t &j);
 
 #endif // INPUT_H
